pure function.
